refactor: Moved auth and login stuff to new location to prepare to move diagrams to new package.
